requirements.txt
openai
beautifulsoup4
requests
python-dotenv

feedparser

beautifulsoup4

streamlit
